+++ src/storybook/Stories.test.ts
@@ -1,9 +1,10 @@
 import createGenerateClassName from "@material-ui/styles/createGenerateClassName";
 import initStoryshots from "@storybook/addon-storyshots";
+import Adapter from "@wojtekmaj/enzyme-adapter-react-17";
 // tslint:disable no-submodule-imports
 import { configure, render } from "enzyme";
-import Adapter from "enzyme-adapter-react-16";
 import toJSON from "enzyme-to-json";
+import { mockRandomForEach } from "jest-mock-random";
 import React from "react";
 
 // Fixes useLayoutEffect warnings
@@ -19,6 +20,10 @@ jest.mock("@material-ui/styles/createGenerateClassName");
 initStoryshots({
   configPath: "src/storybook/",
   test({ story }) {
+    mockRandomForEach(
+      Array.from(Array(25).keys()).map(x => parseFloat(`0.${x + 1}`)),
+    );
+
     const result = render(story.render() as any);
     expect(toJSON(result)).toMatchSnapshot();
   },