+++ CHANGELOG.md
@@ -10,11 +10,13 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix invalid values in channel picker - #2313 by @orzechdev
 - Fix missing metadata and payment balance on unconfirmed orders - #2314 by @orzechdev
 - Fix exit form dialog false positive - #2311 by @orzechdev
+- Enable save button on channel pages - #2328 by @orzechdev
 - Handle form errors before product creation - #2299 by @orzechdev
 - Fix no product error on unconfirmed order lines - #2324 by @orzechdev
 - Enable save button on discount pages - #2319 by @orzechdev
 - Enable save button on page pages - #2325 by @orzechdev
 - Fix pagination errors on voucher and sale pages - #2317 by @orzechdev
+- Add format tip for text attribute rows - #2340 by @orzechdev
 
 ## 3.4